Make font size a factor of viewport height
By making the font size a multiple of the viewport height, we can fit
all the time components on the screen all the time. Yay!

Additionally, because the font-size is variable, the centering we
previously had no longer worked well. To address this, the width of the
time component was fixed, but allowed to overflow. The variance of
the width doesn't look unbalanced when it's "wrong", but because the
width is fixed it doesn't jiggle with that change.

Note the additional `min-height: 0vh` -- this is to work around a safari
bug -- see the link in the change for details.