Persist the theme, and restore it on load
When toggling the theme, you have to toggle it by hand. Thats annoying.

Here we persist it to local storage, and attempt to restore on page load.
This still risks a flat of unthemed content, but it's working OK on my
macbook.

Note, if toggled to 'dark', and then the system toggles to dark, it'll
end up light -- whoops. Problem for another day.