Separate theme override between dark/light
Previously when we toggled the theme, it would toggle to the
'alternative' of whatever theme you were using. If you were light,
toggled to dark, and then switched your system theme to dark, the result
would be a *light* theme, since that was the 'alternative' to dark.

With this change, they're now distinct themes, and persisted distinctly.