Set tab colour explicitly to match the theme
Safari 15 introduced behaviour that detects the colour for the tab bar
from the first pixel of the page. This works great *most* of the time.
But when it doesn't, you need to scroll or render the page to have it
catchup.

This change sets the colour explicitly based on whatever the effective
colour of the body background is.