Refactor clock to use a template
This is a preparatory change to supporting multiple timers. It’s focused
on moving the timer layout out of hard coded HTML in to a template that
is cloned for each countdown itself.

With this change, there is still only a single countdown, with each
countdown having it’s own timer tick. Additionally, shortcuts still
don’t support handling more than one countdown to control.