Made the type of times consistent
There were a few places where we were being terribly inconsistent with
the types we were using for time. Sometimes strings, sometimes date,
sometimes number. On top of that we were being very inconsistent on what
we assumed the types were, and were instantiating new dates from strings
to dates, and then back to strings again.

Whoops. This cleans it to be consistent.