Update default clock selection for upcoming horizon
With the first default date time horizon coming up, I don't want people
to miss out on the celebration, but also have nothing to look forward
to.

This changes the default logic to add up to two countdowns if an
expired countdown is within 5 days of it's target date.

If they're all past, then the 'next jan 1' is the only one seen.

Additionally, *default* clocks are not persisted if there are *only*
default countdowns.