Fix menu sizing on iOS
The sizing was iffy, because `100vh` will take the whole viewport. But
on iOS, this isn't static, and doens't account for the address bar
resizing. Safari _does_ support [`svh`](https://webkit.org/blog/12445/new-webkit-features-in-safari-15-4/) et al.
But other browsers don't; this means it's nice but not helpful.

Rather than tackling that inconcsistently directly, just shrink the max
sizes.