Scroll title to beginning on focus loss
When typing/scrolling in the titel of a countdown, the scroll position
would remain when focus left. This is confusing depending on what the
title was.

If focus leaves the title area, we can scroll to the beginning.

To minimize the confusion if you tab away to look at something, we only
do this if the focus remains in the document