Fix issue with click-drag dismissing dialog
When the mouse button goes down *in* the dialog, and up *out* of the
dialog, the `click` event is still raised.

This leads to weird situation of the dialog dismissing when you least
expect it.

This change handles that by looking at `mousedown` and `mouseup` and
doing our own 'click' detection