Fix focus being set to cause a scroll when the dialog is short
With the 'add' button being the `autofocus` element, focus would jump directly
to that button. In the case of the welcome dialog & small height windows, this
would scroll down, skipping the welcome message. Doh.

Rather than add timing hacks to force the scroll position to the right location,
we made the close button default focus, resulting in the dialog being scrolled
to the top implicitly.