Modularise all the typescript
As preparation for switching the build system (which requires
`isolatedModules` to be enabled), this enables `isolatedModules`
compiler options, and then fixes all the errors. This also includes
conversion of the script tag to be a `module`, and allowing all the
async work that implies -- it's assumed at this time that this delays
loading from a remote source, which will be addressed when the build
system is switched to one that does appropriate code splitting.