Refactor speed to be handled more narrowly than having `goFaster` do the work
As we strive to refactor the segments/intervals to be more efficient, we need to
isolate where we manage the clock tick interval. If we continue to leave it as
it was, changing the interval itself gets stomped by going faster. This might be
OK, but also confusing.

This change isolates the _applied_ tick interval to the `start` method on the
clock. This means that _externally_, we just set the interval and allow our
internal acceleration factor to handle going faster or not.

Specifically:
- `goFaster` no longer sets the interval. It just stops/starts the clock again.
- The acceleration factor goes up by a factor of 3, not 10 every time
- We reduced the delay to the next step after a speed change to reflect the
  time to the next reduced whole interval.