Fix an issue where accelerated time would show negative time & not show the celebration
When using accelerated time, it's possible for a time step to move you past the
end of a countdown. We didn't handle this correctly 'cause we relied on the
countdown itself knowing if it was in the past or not. This wasn't applicable
because we were actually in the future (thanks time travel!). So, we now:
- Clamp remaining time to 0 if its less than 0
- Use a local remaining time check to consider the countdown reached, but *only*
  for display purposes.