Update timer interval to reflect visible segments when segments change
If the smallest unit of time you have visible is minutes, theres no need to run
the clock tick every 1 second -- you can run it once a minute. Or once an hour!

With this change, that is now what happens.

Additionally refactored the `DEFAULT_TICKER_INTERVAL` to be an Enum, and moved
it to `clock.ts`