Add year + month support to countdowns
For countdowns futher than 52 weeks into the future, we just show the total
number of weeks. But, we'd like to be able to show the months, and years.

Having recently adopted the Luxon library for our caculations, we can now easily
show the years & months.

This change adds the elements, and updates logic, to show years & months where
appropriate.