Rework `ThemeManager` to lay the groundwork for more themes
Prior to this change, we only had 'default' and 'toggle' for a colour
scheme. This meant that adding another theme wasn't really possible. We
change `ThemeManager` to use explicit (albeit imperfect) names for
themes, and cycle through those.

More work is required to support true cycling, but this lays the
groundwork.

Additional work includes significant terminology changes (E.g., 'colour
scheme') to make it clearer what is actually going on.