Create `PreferenceKey` to enable center-like alignment
When countdowns are within the viewport, we'd like them be aligned
along their tops *and* overall be centered. But, when placed in the
`ScrollView`, it sizes infinitely. Which is not great. But while it's
possible to finagle certain things to center them *individually*,
getting them aligned is hard.

So we do some _stuff_ to calculate a spacer at the top to ensure they
remain 'centered':
- A preference key (`AlignerKey`) to capture the largest child
(countdown) in the whole list of countdowns
- An Environment item to carry the calculated spacer size down to the
children
- Some overlay `GeometryReaders` to capture the sizes of the items to
do the caculation