Suppress initial layout animation
When the view first loads, it has no `spacerLength`, then layout
happens gaining a value which animates 'cause of the implicit
animation. However, we don't want that first animation -- we want it
for subsequent changes. So, create a new field to hold the 'active'
animation, and *only* set it to non-nil when the click ticks for the
first time