Comment `SegmentView` + minimize `withAnimation`
Primarily, this change comments all the things `SegmentView`. Along the
way it slims down some of the parameters passed in the preview 'cause
they were redundant.

Additionally, due to separate learnings and not wanting to comment it,
we no longer need the `withAnimation` on the toggle to see the
animation when changing the segments value