Clean up initial animation logic
Previously we were detecting first vs second pass and storing an
animation in a state. This seemed messy. Upon understanding
`Transactions`, it became clear there was a better way that wasn't
quite so dependent on things happening in the right way at the right
time.

We can now use the `transaction` modifier at the top level of the app
to set `disableAnimations` based on if we've moved from our initial
tick. Works a treat! We can then respect that deeper in the tree and
do the right thing.

Additionally renamed `now` and `base` to `currentTime` for clarity.