Fix (e.g., make not crashy) suggestions logic
Previously, we had multiple distinct arrays that were processed & merged to be
the final array that was used for suggestions. However, the logic didn't account
for no suggestions for the _next_ year when looking forward from the current
date. This lead to a crash.

While looking at it, it looked like it wasn't the greatest logic in the first
place and could be improved. So, instead of merging multiple arrays, this just
starts with the backstop list as the 'starting' state, and then finds the
additional dates (if there are any). Running out of dates for any year _should_
now result in just the autogenerated year + 2038 (13 years should give enough
time for some more suggestion updates!)