Cycling Segments would restart the clock -- it shouldn't
When cycling segments via `cycleVisibleSegments`, the keyboard handler (`s`)
would automatically restart the clock. This had orignally happened when we were
making the speed up factor respect the time intervals that were _visible_ more
effectively. We were doing it fine, but also forcing the clock to restart, if it
was paused.

With this change, we added a new method (`updateTickInterval`) that:
1. Updates the interval
2. Resets the clock (and starts it ticking) **if** it was already running. If it
   wasn't, it leaves it be.