Switch to `Intl.Segmenter` for grapheme cluster extraction (E.g., better emoji)
The emoji mechanism we were using -- a regex with some matching -- was great,
but it didn't capture all the fancy ZWJ combined emoji's that would compose over
multiple emojis and collapse to a single composite emoji. You'd get confetti of
each constituent emoji. Le Sad.

This version uses `Intl.Segmenter` which knows what it's doing to extract them
from the supplied string, and then just a simple `\p{Extended_Pictograph}` to
match if it's got an actual emoji in it.

Also:
- add these suspect emojis to the sample test data
- make the debug launch automatically include the `testtimes=true` for ease