Adjust Title bar min size + fix scroll bar overlaying
The title of the countdowns had the `min-width` property removed a while ago as
part of some other tweaks. We missed that, uhh, that means the title container
(which is used to give scrolling powers) would collapse to 0 'cause it has no
real content to size it. This meant all the items would smoosh together. Putting
a min width of ~1.2ch on this fixed it.

While doing this, I was also reminded about the terrible scrollbar that pops up
when you scroll the actual title (assuming it's too long to fit) you might see a
huge scrollbar covering your text. Euw. So, to fix this:
1. Introduced another level of indirection element to contain the item that
   will be the scroller
2. Made the scroller overflow the indirection element just added.
3. Replaced the box shadow with an `::after` pseudo element sized to the
   indirection element, and then futzed with the size to feel right.

Magic.